+++ NEWS.md
@@ -22,8 +22,9 @@
 * Extend `read.ctd.sbe()` to handle 4 new (Beckman) oxygen variables.
 * Extend `read.ctd.sbe()` to handle `.btl` files.
 * Extend `oce.plot.ts()`, adding argument `logStyle`.
-* Change user-oriented github website generator to `pkgdown`.
 * Fix `plot,coastline-method` error in box drawing.
+* Fix `subset,argo-method()` error in `flags`, `location`, and `*QC` in `metadata`.
+* Change user-oriented github website generator to `pkgdown`.
 
 ## 1.2.0